% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linguistic_distance.R
\name{calc_lang_dist}
\alias{calc_lang_dist}
\title{Calculate linguistic distance}
\usage{
calc_lang_dist(ethno.g, vertex.a, vertex.b, expand = F, delta = 0.5)
}
\description{
Calculate linguistic distance
}
\keyword{internal}
