% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tree_link.R
\name{check_lang_links}
\alias{check_lang_links}
\title{Check hand-coded links from ethnic groups to languages}
\usage{
check_lang_links(link.df, ethno.long)
}
\description{
Check hand-coded links from ethnic groups to languages
}
\keyword{internal}
