% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tree_link.R
\name{consistent_lang_link}
\alias{consistent_lang_link}
\title{Wrapper to link ethnic groups to language tree}
\usage{
consistent_lang_link(link.df, ethno.long, list.dict)
}
\description{
Wrapper to link ethnic groups to language tree
}
\keyword{internal}
