% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coder_link_lang.R
\name{fuzzy_str_match}
\alias{fuzzy_str_match}
\title{Fuzzy Matching of string to target string}
\usage{
fuzzy_str_match(
  source,
  target,
  target.name = target,
  target.type,
  threshold = 0.2,
  costs = c(insertions = 1, deletions = 1, substitutions = 1)
)
}
\description{
Fuzzy Matching of string to target string
}
\keyword{internal}
