% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linguistic_distance.R
\name{ling_dist_df}
\alias{ling_dist_df}
\title{Make linguistic distance DataFrame}
\usage{
ling_dist_df(
  link.a,
  link.b,
  link.dict,
  list.dict,
  level = c("dialect", "langauge"),
  by.country = T,
  delta = 0.5,
  expand = FALSE,
  agg_fun.a = min,
  agg_fun.b = min,
  ethno.df,
  ethno.long,
  ethno.g = NULL
)
}
\description{
Make linguistic distance DataFrame
}
\keyword{internal}
