% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_ethnonode_set.R
\name{link_any}
\alias{link_any}
\title{Set link, work horse function}
\usage{
link_any(
  link.a,
  link.b,
  link.level,
  link.on.country = T,
  minimalist = F,
  ethno.same.country = F,
  link.dict,
  list.dict,
  drop.a.threshold = 0,
  drop.b.threshold = 0
)
}
\description{
Set link, work horse function
}
\note{
Note that, for coding reasons,
  'dialects' are level 17 in our language dictionary.
  Levels 15 and 16 are languages, identified via their
  name (level 15) and iso code (level 16).
}
\keyword{internal}
