% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tree_link.R
\name{link_long_4steps}
\alias{link_long_4steps}
\title{Main function to link ethnic groups to language tree}
\usage{
link_long_4steps(ethno.long, link.long)
}
\description{
Main function to link ethnic groups to language tree
}
\keyword{internal}
