% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tree_link.R
\name{links_wide2long}
\alias{links_wide2long}
\title{Function to transform handcoded links of ethnic groups to languages from wide to long format.}
\usage{
links_wide2long(link.df)
}
\description{
Returns data so that each link has its own row, adds a variable for the level of the link
}
\keyword{internal}
