% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linguistic_distance.R
\name{make_ling_tree}
\alias{make_ling_tree}
\title{Transform wide language dataset into igraph tree}
\usage{
make_ling_tree(df, collapse = T)
}
\description{
Transform wide language dataset into igraph tree
}
\keyword{internal}
